backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
face-recognition==1.3.0
pillow==10.1.0
bcrypt==4.0.1
python-jose==3.3.0